tests/wpt/web-platform-tests/websockets/constructor/018.html

<!doctype html>
<title>WebSockets: NULL char in url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo-query?x\u0000');
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, 'x%00');
    ws.close();
    t.done();
  })
  ws.onclose = ws.onerror = t.step_func(function(e) {assert_unreached(e.type)});
});
</script>
